requirements.txt
icalendar~=5.0.4
paramiko~=3.0.0
keyring~=23.13.1
click~=8.1.3
python-dotenv~=1.0.0
python-dateutil~=2.8.2
